pyproject.toml
[tool.poetry]
name = "bedrock_deep_search"
version = "0.1.0"
description = "Bedrock Deep Search"
authors = ["Lorenzo Micheli <miclor@amazon.com>, Akarsha Sehwag <akshseh@amazon.de>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
langgraph = "^0.2.44"
langchain-aws = "^0.2.6"
langchain = "^0.3.7"
python-dotenv = "^1.0.1"
langgraph-checkpoint-sqlite = "^2.0.1"
tavily-python = "^0.5.0"
pydantic = "^2.9.2"
watchdog = "^6.0.0"
beautifulsoup4 = "^4.12.3"
langchain-community = "^0.3.8"
streamlit = "1.42.2"
pyperclip = "^1.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
ipykernel = "^6.29.5"
pytest-cov = "^6.0.0"
bandit = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
